Work on Part Two of the Next.js blog tutorial
Finished the second part of the Next.js blog tutorial. This dealt mostly
with working with CSS (SCSS modules in my case), metadata, and asset
handling. In addition, lots of work went into fine tuning the dev setup
with ESLint so that it works the way I want it to now.